refactor(test/fixture/): Fixture 클래스를 모아둔 패키지 생성
